basic/requirements.txt
tensorflow-gpu
numpy
pandas
scikit-learn
math
re
os
IPython
tensorflow-gpu


                                                                                                 
